weather-pipeline/requirements.txt
apache-airflow~=2.5.3
apache-airflow-providers-http==4.0.0
pandas>=1.3.0,<1.4.0
requests>=2.25.0,<2.29.0
psycopg2-binary>=2.9.0
sqlalchemy<2.0
Flask~=2.2.5